requirements.txt
clashai
python-dotenv
pyTelegramBotAPI